refactor(W3D2): I/O 및 Application 관련 클래스 Application 패키지로 분리
- application, domain, infrastructure 구조로 패키지 재설계
