+++ kdt-spring-order/src/main/java/org/prgrms/kdt/application/Console.java
@@ -1,0 +1,8 @@
+package org.prgrms.kdt.application;
+
+import org.prgrms.kdt.application.voucher.io.Input;
+import org.prgrms.kdt.application.voucher.io.Output;
+
+public class Console implements Input, Output {
+
+}
