+++ kdt-spring-order/src/main/java/org/prgrms/kdt/application/voucher/VoucherApplication.java
@@ -1,15 +1,15 @@
-package org.prgrms.kdt.domain.voucher.application;
+package org.prgrms.kdt.application.voucher;
 
-import org.prgrms.kdt.domain.voucher.io.Input;
-import org.prgrms.kdt.domain.voucher.io.Output;
+import org.prgrms.kdt.application.voucher.io.Input;
+import org.prgrms.kdt.application.voucher.io.Output;
 import org.prgrms.kdt.domain.voucher.service.VoucherService;
 
-public class VoucherCommandLineApplication implements Runnable{
+public class VoucherApplication implements Runnable{
     private VoucherService voucherService;
     private Input input;
     private Output output;
 
-    public VoucherCommandLineApplication(VoucherService voucherService, Input input, Output output) {
+    public VoucherApplication(VoucherService voucherService, Input input, Output output) {
         this.voucherService = voucherService;
         this.input = input;
         this.output = output;
