+++ kdt-spring-order/src/main/java/org/prgrms/kdt/application/voucher/io/Output.java
@@ -1,0 +1,4 @@
+package org.prgrms.kdt.application.voucher.io;
+
+public interface Output {
+}
