--- kdt-spring-order/src/main/java/org/prgrms/kdt/domain/voucher/io/Input.java
@@ -1,4 +1,0 @@
-package org.prgrms.kdt.domain.voucher.io;
-
-public interface Input {
-}
