--- kdt-spring-order/src/main/java/org/prgrms/kdt/infrastructure/voucher/io/Console.java
@@ -1,8 +1,0 @@
-package org.prgrms.kdt.infrastructure.voucher.io;
-
-import org.prgrms.kdt.domain.voucher.io.Input;
-import org.prgrms.kdt.domain.voucher.io.Output;
-
-public class Console implements Input, Output {
-
-}
